+++ common/persistence/sql/sqlplugin/definition.go
@@ -99,7 +99,7 @@ type (
 
 	// ShardsRow represents a row in shards table
 	ShardsRow struct {
-		ShardID      int64
+		ShardID      int32
 		RangeID      int64
 		Data         []byte
 		DataEncoding string
@@ -108,12 +108,12 @@ type (
 	// ShardsFilter contains the column names within shards table that
 	// can be used to filter results through a WHERE clause
 	ShardsFilter struct {
-		ShardID int64
+		ShardID int32
 	}
 
 	// ExecutionsRow represents a row in executions table
 	ExecutionsRow struct {
-		ShardID                  int
+		ShardID                  int32
 		NamespaceID              primitives.UUID
 		WorkflowID               string
 		RunID                    primitives.UUID
@@ -130,7 +130,7 @@ type (
 	// ExecutionsFilter contains the column names within executions table that
 	// can be used to filter results through a WHERE clause
 	ExecutionsFilter struct {
-		ShardID     int
+		ShardID     int32
 		NamespaceID primitives.UUID
 		WorkflowID  string
 		RunID       primitives.UUID
@@ -138,7 +138,7 @@ type (
 
 	// CurrentExecutionsRow represents a row in current_executions table
 	CurrentExecutionsRow struct {
-		ShardID          int64
+		ShardID          int32
 		NamespaceID      primitives.UUID
 		WorkflowID       string
 		RunID            primitives.UUID
@@ -152,7 +152,7 @@ type (
 	// CurrentExecutionsFilter contains the column names within current_executions table that
 	// can be used to filter results through a WHERE clause
 	CurrentExecutionsFilter struct {
-		ShardID     int64
+		ShardID     int32
 		NamespaceID primitives.UUID
 		WorkflowID  string
 		RunID       primitives.UUID
@@ -231,7 +231,7 @@ type (
 	// ReplicationTaskDLQRow represents a row in replication_tasks_dlq table
 	ReplicationTaskDLQRow struct {
 		SourceClusterName string
-		ShardID           int
+		ShardID           int32
 		TaskID            int64
 		Data              []byte
 		DataEncoding      string
@@ -281,7 +281,7 @@ type (
 
 	// HistoryNodeRow represents a row in history_node table
 	HistoryNodeRow struct {
-		ShardID  int
+		ShardID  int32
 		TreeID   primitives.UUID
 		BranchID primitives.UUID
 		NodeID   int64
@@ -294,7 +294,7 @@ type (
 	// HistoryNodeFilter contains the column names within history_node table that
 	// can be used to filter results through a WHERE clause
 	HistoryNodeFilter struct {
-		ShardID  int
+		ShardID  int32
 		TreeID   primitives.UUID
 		BranchID primitives.UUID
 		// Inclusive
@@ -306,7 +306,7 @@ type (
 
 	// HistoryTreeRow represents a row in history_tree table
 	HistoryTreeRow struct {
-		ShardID      int
+		ShardID      int32
 		TreeID       primitives.UUID
 		BranchID     primitives.UUID
 		Data         []byte
@@ -316,7 +316,7 @@ type (
 	// HistoryTreeFilter contains the column names within history_tree table that
 	// can be used to filter results through a WHERE clause
 	HistoryTreeFilter struct {
-		ShardID  int
+		ShardID  int32
 		TreeID   primitives.UUID
 		BranchID primitives.UUID
 	}
